arm64: Add support for configuring PMU counters
Configure the number of PMU counters of a VM by reading and writing the
PMCR_EL0 register with KVM_GET/SET_ONE_REG.

Use KVM_GET/SET_ONE_REG to configure the PMU counters. Use '-1' as
default value, so we can detect when the user requests the number of
counters on the command-line (0 is a valid number of counters).

Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>