Add TPM Event log
Verification of a confidential VM is easier with an event log. When Add
a tool to generate a TPM event log, following the TCG TPM2 standard.
Events such as loading an image into the VM can be added to the log,
along with a hash of the image, that a verifier can retrieve when trying
to recreate the initial VM measurements.

To compute the fingerprint of images, add a dependency on libcrypto
(OpenSSL).

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>